修改订单

+++ src/main/java/com/qianmo/eshop/resource/seller/OrderResource.java
@@ -89,7 +89,7 @@ public class OrderResource extends SellerResource {
                 buyerUsername = buyerUser.get("username");
             }
             String message = "订单"+order_num+"已发货，发货人："+userInfo.get("nickname")+"，联系电话："+userInfo.get("phone");
-            Integer orderId = order_info.dao.findFirstBy("num = ?", order_num).get("id");
+            Long orderId = order_info.dao.findFirstBy("num = ?", order_num).get("id");
             JPushClientServer.pushMassage(buyerUsername,message,String.valueOf(orderId));
             new order_remark().set("order_num", order_num).set("op", op).set("details", remark == null ? "" : remark).set("area_id", ConstantsUtils.ALL_AREA_ID).set("user_id", orderUser.get("seller_id")).save();
             sms_template model = sms_template.dao.findById(ConstantsUtils.INVITE_VERIFY_CODE_TYPE_SEND_SMS);
